Move the paths module
