+++ src/amulet/app/path/__init__.py
@@ -1,0 +1,1 @@
+from ._application import data_directory, config_directory, cache_directory, logging_directory
