+++ src/amulet/app/path/_plugin.py
