--- src/amulet_editor/data/paths/__init__.py
@@ -1,5 +1,0 @@
-from amulet_editor.data.paths._application import (
-    data_directory,
-    project_directory,
-    user_directory,
-)
