+++ src/amulet_editor/data/plugin/_manager.py
@@ -28,7 +28,7 @@ from PySide6.QtCore import Signal, QObject
 from amulet_editor.application._splash import Splash
 from amulet_editor.application._invoke import invoke
 
-from amulet_editor.data.paths._plugin import (
+from amulet.app.path._plugin import (
     first_party_plugin_directory,
     third_party_plugin_directory,
 )
